+++ releases.md
@@ -5,6 +5,7 @@
   - Add `Capture#each_object_id` for getting object identifiers.
   - Add `retained_addresses:` option to `Sampler#analyze` to capture addresses.
   - Add `Sampler#analyze(retained_minimum: 100)` - if the retained_size is less than this, the analyse won't proceed.
+  - Remove `Memory::Profiler::Graph` - it's too slow for practical use.
 
 ## v1.4.0